+++ front-of-noise/src/App.js
@@ -1,7 +1,6 @@
 import React, { Component } from 'react';
 // import {Helmet} from 'react-helmet';
-import 'bulma';
-import './App.css';
+import './styles/css/App.css';
 import Layout from './app/Layout'
 
 class App extends Component {
